Move internal mutex interface into environment; Clean up docs
